+++ packages/lint-worker/src/helpers.ts
@@ -7,28 +7,26 @@ import {
 import axios from 'axios';
 import { DbSchema as DbSchemaV1 } from 'languageSupport-next.13';
 
-const oldLinter = '5.20.0';
-
 // for older versions of the language support, the dbschema was not the same,
 // meaning old linters need conversion of the new schema
 export function convertDbSchema(
   originalSchema: DbSchemaV2,
   linterVersion: string,
 ): DbSchemaV2 | DbSchemaV1 {
-  let oldFunctions: Record<string, Neo4jFunction> = {};
-  let oldProcedures: Record<string, Neo4jProcedure> = {};
   if (!originalSchema) {
     return originalSchema;
   }
-  if (
-    originalSchema.functions['CYPHER 5'] &&
-    originalSchema.procedures['CYPHER 5']
-  ) {
-    oldFunctions = originalSchema.functions['CYPHER 5'];
-    oldProcedures = originalSchema.procedures['CYPHER 5'];
-  }
 
-  if (compareMajorMinorVersions(linterVersion, oldLinter) <= 0) {
+  if (compareMajorMinorVersions(linterVersion, '2025.01') < 0) {
+    let oldFunctions: Record<string, Neo4jFunction> | undefined = undefined;
+    let oldProcedures: Record<string, Neo4jProcedure> | undefined = undefined;
+    if (originalSchema.functions && originalSchema.functions['CYPHER 5']) {
+      oldFunctions = originalSchema.functions['CYPHER 5'];
+    }
+
+    if (originalSchema.procedures && originalSchema.procedures['CYPHER 5']) {
+      oldProcedures = originalSchema.procedures['CYPHER 5'];
+    }
     const dbSchemaOld: DbSchemaV1 = {
       ...originalSchema,
       functions: oldFunctions,
@@ -40,12 +38,30 @@ export function convertDbSchema(
   }
 }
 
-export function serverVersionToLinter(serverVersion: string) {
-  let candidate: string = 'Latest';
-  if (compareMajorMinorVersions(serverVersion, oldLinter) <= 0) {
-    candidate = oldLinter;
+export function serverVersionToLinter(serverVersion: string): {
+  linterVersion: string;
+  notResolved?: boolean;
+  notSupported?: boolean;
+} {
+  // Extract only the major and minor
+  const versionRegex = /^\d+\.\d+/;
+  const linterVersion = serverVersion.match(versionRegex)?.[0];
+
+  // If we have a version lower than 5.23, use that linter
+  if (compareMajorMinorVersions(serverVersion, '5.23') < 0) {
+    return { linterVersion: '5.23', notSupported: true };
+    // Unfortunately 2025.01, 2025.02 and 2025.03 all return 5.27
+    // so we have to assume we are on the most modern database from all those
+    //
+    // This case should never happen though because we should have cleaned up the
+    // version by the moment we call this method
+  } else if (compareMajorMinorVersions(serverVersion, '5.27') === 0) {
+    return { linterVersion: '2025.03' };
+  } else if (linterVersion) {
+    return { linterVersion: linterVersion };
   }
-  return candidate;
+
+  return { linterVersion: 'Default', notResolved: true };
 }
 
 export function linterFileToServerVersion(fileName: string) {
@@ -64,7 +80,7 @@ export type NpmRelease = {
 };
 
 export const npmTagToLinterVersion = (tag: string) =>
-  tag.match(/^neo4j-([\d.]+)$/)?.[1];
+  tag.match(/^neo4j-(\d+\.\d+)$/)?.[1];
 
 export async function getTaggedRegistryVersions(): Promise<NpmRelease[]> {
   const registryUrl = 'https://registry.npmjs.org/@neo4j-cypher/lint-worker';
@@ -74,7 +90,7 @@ export async function getTaggedRegistryVersions(): Promise<NpmRelease[]> {
     const taggedVersions: { tag: string; version: string }[] = [];
     if (data !== null && data['dist-tags'] !== null) {
       for (const [tag, version] of Object.entries(data['dist-tags'])) {
-        if (typeof tag === 'string' && typeof version === 'string') {
+        if (npmTagToLinterVersion(tag)) {
           taggedVersions.push({ tag, version });
         }
       }